add logger and fix bug of init

+++ cmd/generate.go
@@ -1,19 +1,45 @@
 package cmd
 
 import (
+	"errors"
 	"fmt"
-	"github.com/pelletier/go-toml/v2"
+	"forge/internal/devices"
+	"forge/internal/logger"
+	"log"
 	"os"
 	"path/filepath"
 	"strings"
+
+	"github.com/pelletier/go-toml/v2"
 )
 
-func validateDevice(device string) error {
-	if !strings.HasPrefix(strings.ToUpper(device), "STM32") {
-		return fmt.Errorf("invalid device: %s", device)
+func updateProjectData(cfg *tomlConfig) (projectData, error) {
+	pd := projectData{
+		Name:      cfg.Project.Name,
+		Version:   cfg.Project.Version,
+		BuildType: cfg.Build.Type,
 	}
 
-	return nil
+	switch {
+	case cfg.Target.Kind == "host-x64":
+		pd.Architecture = "x86_64"
+		pd.Compiler = compilersMap["x86_64"]
+		pd.Target = nil
+	case cfg.Target.Kind == "mcu" && strings.HasPrefix(strings.ToUpper(cfg.Target.Device), "STM32"):
+		pd.Architecture = "cortexM"
+		pd.Compiler = compilersMap["CortexM"]
+		pd.Target = &targetData{
+			Device: cfg.Target.Device,
+			// CPU:    stm32CPUMap[strings.ToUpper(cfg.Target.Device)],
+			Vendor: "STMicroelectronics",
+			Family: "STM32",
+			Series: strings.ToUpper(cfg.Target.Device),
+		}
+	default:
+		return pd, fmt.Errorf("invalid target kind")
+	}
+
+	return pd, nil
 }
 
 func New(args ...string) error {
@@ -21,85 +47,90 @@ func New(args ...string) error {
 	setConfigDefaults()
 	nameProject := args[0]
 	setConfig(nameProject)
+	msgErr := fmt.Errorf("Falied to generate new project.")
 
-	fmt.Println("Initializing a new project:", nameProject)
+	logger.Infof("Initializing a new project:", nameProject)
 
 	_, err := os.Stat(nameProject)
 
 	// If no error, the path already exists
 	if err == nil {
-		logError.Println("Project directory already exists.")
-		return fmt.Errorf("project directory already exists")
+		logger.Error("Project directory already exists.")
+		return err
 	}
 
 	// If we got an error that is not "not exist", return it
 	if err != nil && !os.IsNotExist(err) {
-		logError.Println("Failed to check project directory:", err)
-		return err
+		logger.Error(err)
+		return msgErr
 	}
 
 	err = os.MkdirAll(nameProject, 0755)
 	if err != nil {
-		logError.Println("Failed to create project directory.", err)
-		return err
+		logger.Error(err)
+		return msgErr
 	}
-
-	forgeTOMLName := "Forge.toml"
 
 	if args[1] == "--device" {
 		if len(args) > 2 {
 			device := args[2]
-			err := validateDevice(device)
+			catalog, err := devices.Lookup(device)
 			if err != nil {
-				logError.Println("Invalid device specified. ")
-				return err
+				logger.Error(err)
+				return msgErr
 			}
-			fmt.Println("Target device set to:", device)
 
-			config.Target.Device = device
-			config.Toolchain.Compiler = "gcc-arm-none-eabi"
+			config.Target.Device = catalog.ID
+			config.Target.Kind = "mcu"
 		} else {
-			logError.Println("Device not specified. Use --device <device> to specify the target device.")
-			return fmt.Errorf("device not specified")
+			logger.Error("Device not specified. Use --device <device> to specify the target device.")
+			log.Println("To list supported devices, run 'forge list'.")
+			return msgErr
 		}
 	}
 	_, err = os.Create(filepath.Join(config.Project.Name, forgeTOMLName))
 	if err != nil {
-		logError.Println("Failed to create ", forgeTOMLName, " file.", err)
-		return err
+		logger.Error(err)
+		return msgErr
 	}
 
 	data, err := toml.Marshal(config)
 	if err != nil {
-		logError.Fatal(err)
+		logger.Error(err)
+		return msgErr
 	}
 	err = os.WriteFile(filepath.Join(config.Project.Name, forgeTOMLName), data, 0644)
 	if err != nil {
-		logError.Println("Failed to write to ", forgeTOMLName, " file.", err)
-		return err
+		logger.Error(err)
+		return msgErr
 	}
 	return nil
 }
 
 func Init() error {
-
+	msgErr := fmt.Errorf("Failed to initializing project.")
+	_, err := os.Stat(forgeTOMLName)
+	if errors.Is(err, os.ErrNotExist) {
+		logger.Error(err)
+		return msgErr
+	}
 	for _, nameFolder := range listFolders {
 		err := os.MkdirAll(filepath.Join(config.Project.Name, nameFolder), 0755)
 		if err != nil {
-			logError.Println("Failed to create folder:", nameFolder, err)
-			return err
+			logger.Error("Failed to create folder:", nameFolder, err)
+			return msgErr
 		}
 	}
 
 	for name, Content := range listFilesContentMap {
 		_, err := os.Create(filepath.Join(config.Project.Name, name))
 		if err != nil {
-			logError.Println("Failed to create ", name, " file.", err)
+			logger.Error("Failed to create ", name, " file.", err)
 			return err
 		}
 		err = os.WriteFile(filepath.Join(config.Project.Name, name), []byte(Content), 0644)
 		if err != nil {
-			logError.Println("Failed to write to ", name, " file.", err)
+			logger.Error("Failed to write to ", name, " file.", err)
 			return err
 		}
 	}
